Uploading Projects

+++ Project-1/src/main/java/Main.java
@@ -11,29 +11,29 @@ public class Main {
         }).start(9000);
 
         app.post("/form-add", context -> {
-            int num1 = Integer.parseInt(context.formParam("add"));
-            int num2 = Integer.parseInt(context.formParam("add2"));
+            double num1 = Double.parseDouble(context.formParam("add"));
+            double num2 = Double.parseDouble(context.formParam("add2"));
 
             context.result(String.valueOf((num1+num2)));
         });
 
         app.post("/form-subtract", context -> {
-            int num1 = Integer.parseInt(context.formParam("subtract"));
-            int num2 = Integer.parseInt(context.formParam("subtract2"));
+            double num1 = Double.parseDouble(context.formParam("subtract"));
+            double num2 = Double.parseDouble(context.formParam("subtract2"));
 
             context.result(String.valueOf((num1-num2)));
         });
 
         app.post("/form-multiply", context -> {
-            int num1 = Integer.parseInt(context.formParam("multiply"));
-            int num2 = Integer.parseInt(context.formParam("multiply2"));
+            double num1 = Double.parseDouble(context.formParam("multiply"));
+            double num2 = Double.parseDouble(context.formParam("multiply2"));
 
             context.result(String.valueOf((num1*num2)));
         });
 
         app.post("/form-divide", context -> {
-            int num1 = Integer.parseInt(context.formParam("divide"));
-            int num2 = Integer.parseInt(context.formParam("divide2"));
+            double num1 = Double.parseDouble(context.formParam("divide"));
+            double num2 = Double.parseDouble(context.formParam("divide2"));
 
             context.result(String.valueOf((num1/num2)));
         });
